move the test setup in `setup.rs`
