+++ tests/parallel/main.rs
@@ -1,1 +1,3 @@
+mod setup;
+
 mod cancellation;
